// modules/frontend/src/modules/core/services/generated/model/angleRelation.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * Tu portal al conocimiento y al éxito profesional.
 * OpenAPI spec version: v1
 */
import type { AngleRelationAngleClass } from './angleRelationAngleClass';

export interface AngleRelation {
  angleClass?: AngleRelationAngleClass;
  /**
   * @minLength 0
   * @maxLength 200
   */
  detail?: string;
}