// modules/frontend/src/modules/core/services/generated/model/midline.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */
import type { MidlinePosition } from './midlinePosition';

export interface Midline {
  position?: MidlinePosition;
  /** @maximum 30 */
  deviationMm?: number;
  deviated?: boolean;
}