// modules/frontend/src/modules/core/services/generated/model/sideRelations.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * Tu portal al conocimiento y al éxito profesional.
 * OpenAPI spec version: v1
 */
import type { AngleRelation } from './angleRelation';

export interface SideRelations {
  right?: AngleRelation;
  left?: AngleRelation;
}